Pause/Resume on prediction.

+++ src/UI/emotion_recognition/VoiceEmotionDetectionThread.py
@@ -26,12 +26,15 @@ class VoiceEmotionDetectionThread(QThread):
         self._pyAudioObject = pyaudio.PyAudio()
         self._audio_input_stream = None
 
-        self._frames = []
+        self._frames_to_predict = []
         self._predicted_frames = []
+        self._frames = []
         self._manager = Manager()
         self._emotion = {0: 'Angry', 1: 'Disgust', 2: 'Fear', 3: 'Happy', 4: 'Neutral', 5: 'Sad', 6: 'Surprise'}
         self._chunk_step = 16000
         self._chunk_size = 49100
+
+        self._is_paused = False
 
     def frame(self, y, win_step=64, win_size=128):
         # Number of frames
@@ -60,6 +63,7 @@ class VoiceEmotionDetectionThread(QThread):
         return np.asarray(mel_spect)
 
     def run(self):
+        self._is_paused = False
         self._audio_input_stream = self._pyAudioObject.open(
             format=pyaudio.paInt16,
             channels=self._channels,
@@ -74,25 +78,35 @@ class VoiceEmotionDetectionThread(QThread):
             while self._audio_input_stream.is_active():
                 data = self._audio_input_stream.read(self._frames_per_buffer)
                 self._frames.append(data)
+
+                if self._is_paused:
+                    continue
+
+                self._frames_to_predict.append(data)
                 end_time = time.time()
                 seconds_passed = end_time - start_time
                 if seconds_passed > 4:
-                    start_time = time.time()
                     # data = wave_utils.convert_to_wave(self._frames)
+                    if not self._is_paused:
+                        # Alternative method until I fix the stuff with reading from byte class
+                        file_name = "./Temp/" + str(uuid.uuid4()) + '.wav'
+                        wave_utils.write_wave(file_name, self._frames_to_predict)
+                        data, _ = wave_utils.load_wave(file_name)
+                        os.remove(file_name)
 
-                    # Alternative method until I fix the stuff with reading from byte class
-                    file_name = "./Temp/" + str(uuid.uuid4()) + '.wav'
-                    wave_utils.write_wave(file_name, self._frames)
-                    data, _ = wave_utils.load_wave(file_name)
-                    os.remove(file_name)
+                        prediction = self.predict_audio(data)[0]
+                        str_prediction = f"Current voice emotion detect as: {prediction}"
+                        self._parent.chart.setTitle(str_prediction)
+                        print(str_prediction)
 
-                    prediction = self.predict_audio(data)[0]
-                    str_prediction = f"Current voice emotion detect as: {prediction}"
-                    self._parent.chart.setTitle(str_prediction)
+                        self._predicted_frames.append((self._frames_to_predict, prediction))
 
-                    print(str_prediction)
-                    self._predicted_frames.append((self._frames, prediction))
-                    self._frames.clear()
+                    self._frames_to_predict.clear()
+                    start_time = time.time()
+
+            wave_utils.write_wave("Candidate_Audio.wav", self._frames)
+            self._frames.clear()
+
         except Exception as ex:
             self._logger.log_error(ex)
             raise Exception(ex)
@@ -131,6 +145,13 @@ class VoiceEmotionDetectionThread(QThread):
         predict = [self._emotion.get(emotion) for emotion in predict]
         return predict
 
-    def stop_recording(self):
+    def stop_prediction(self):
         self._audio_input_stream.stop_stream()
         self._audio_input_stream.close()
+
+    def pause_prediction(self):
+        self._is_paused = True
+        self._parent.chart.setTitle("Prediction is paused ")
+
+    def resume_prediction(self):
+        self._is_paused = False
